Tiny changes

+++ createidx.h
@@ -15,7 +15,7 @@
 #include "util.h"
 
 // Functions
-void createIDX( const std::string & wordlist, const std::string & idxFile, const std::string & hash, long cores, bool quiet );
+void createIDX( const std::string & wordlist, const std::string & idxFile, const std::string & hash, size_t cores, bool quiet );
 void computeHashes( std::atomic<bool>* threadReady, std::mutex* fileInMutex, std::mutex* fileOutMutex, std::ifstream* fileIn, std::ofstream* fileOut );
 
 #endif
